backend/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pymongo
langchain_sambanova
fitz
python-docx
shutil
